pyproject.toml
[project]
name = "mdnews"
version = "0.1.0"
description = ""
authors = [{ name = "Iván SZKIBA", email = "szkiba@users.noreply.github.com" }]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "rss-parser (>=2.1.0,<3.0.0)",
    "markitdown (>=0.0.1a3,<0.0.2)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
